Return or set column_header_buffer attributes

These property functions modify and return the column header buffers of a rtf_doc object. These are stored as a named vector. Names should be top and bottom. This attribute adds rows to the top or bottom of the table column headers to pad it from the titles above or the table below.

column_header_buffer(x, ...)

column_header_buffer(x) <- value

set_column_header_buffer(x, ...)

Arguments

x
A rtf_doc object

...
Additonal arguments passed to method dispatch. Should include argument top and bottom with numeric elements.

value
A named vector detailing the top and bottom buffer.

Value

For column_header_buffer, the column_header_buffer attribute of the supplied rtf_doc. For `column_header_buffer<-`() and set_column_header_buffer(), the modified object.

Examples

library(huxtable)
ht <- huxtable(
  column1 = 1:5,
  column2 = letters[1:5]
)
rtf <- rtf_doc(ht, list(hf_line("aTitle")))

column_header_buffer(rtf)

#>    top bottom 
#>      0      0 

# Returns c(top = 0, bottom = 0), the default

column_header_buffer(rtf) <- c(bottom = 1)
# Sets the bottom column_header_buffer to 1